Add configuration
